purgo_ai/requirements.txt
# 의존성

flask
openai
python-dotenv
